🔨 chore: refactor MangaList model import and remove duplicate code
The import statement for the MangaList model has been refactored to import it from the `app.functions.class_mangalist` module. This helps to improve code organization and avoid duplicate code.
🔄 refactor: remove duplicate import and unused class

- Remove duplicate import statement for `declarative_base` from `sqlalchemy.orm`.
- Remove unused `MockManga` class definition,
import from file
🔧 fix: import declarative_base from sqlalchemy.orm instead of sqlalchemy
